Bring TaxonomyPicker sample and Component in sync
Update sample and component to be in sync. In order to do
this I reimplemented the Samples functionality to accept
context, callbacks, LevelToShowTerms and Filter.

There is an option in the sample to use term as root node which
I did not reimplement, I don't really see the point of it and
it's not really trivial to do.

I did not update the README's of the projects, it warrants
a discussion of why there are two (now in sync) things in PnP
that does the same thing. Why not remove the sample?