emails containing apostrophe fix
We recently had an issue at a customer where an email address contained an apostrophe. According to the spec RFC 2822 (https://tools.ietf.org/html/rfc2822) This is actually possible. Up until now I was not aware of this. 
Without this fix, we could not select a user with an apostrophe in the email address from the people picker.