Provisioning.pn p deployer.console (#1552)
* Added a configurable parameter that allows the deployer to ignore folders in order to speed up the tokenizing process. Updated the token file XSD for a more consistent XML. Fixed a bug related to tokens loaded from the App.config file. Removed the project build events.

* Update FieldsExtensibilityProvider.md

* Update README.md

* Added the custom Required attribute to the FieldsExtensibilityProvider serializable classes.

* Added the ignored extensions functionnality

Added the ignored extensions functionnality in order to be able to ignore specific file extensions for the tokenizing process. This makes sure that files such as images, videos, pnp packages etc aren't corrupted by the tokenizer opening them and saving them in a corrupted format.

* Added the ignored extensions functionnality

Added the ignored extensions functionality in order to be able to ignore specific file extensions for the tokenizing process. This makes sure that files such as images, videos, pnp packages etc aren't corrupted by the tokenizer opening them and saving them in a corrupted format.

* Added the ignored extensions functionnality

Added the ignored extensions functionnality in order to be able to ignore specific file extensions for the tokenizing process. This makes sure that files such as images, videos, pnp packages etc aren't corrupted by the tokenizer opening them and saving them in a corrupted format.

* Removed the previously added ignored extensions parameter

Removed the previously added ignored extensions parameter and replaced it by a more dynamic way of handling the problem.

* Removed the previously added ignored extensions parameter

Removed the previously added ignored extensions parameter and replaced it by a more dynamic way of handling the problem.

* Removed the previously added ignored extensions parameter

Removed the previously added ignored extensions parameter and replaced it by a more dynamic way of handling the problem.